export section and write as global option
Some YaST programs queries the bootloader's section information to
manipulate it's settings, for example the yast2-kdump.

The pbl will now read grub.cfg and extract each section's information
in order to export to yast. Yast programs now shouldn't fail due to
lack of these information. And when it attempts to modify, for example
the section's kernel's parameter, it will be directed to the
corresponding global option. (bnc#773143)