handle add|remove in update-bootloader for grub2
Per man page of update-bootloader, it says add or remove need to call
the refresh to take effect, but in fact it's invalid for grub or grub2
since they do not require refresh (reinstall) to get new config take
effect. Thus we'd like to change the grub2 to follow this.

1. If refresh was specified, will update config and re-install
2. If add or remove specified, update the config only

This change also verified to fix "Menu entry for memtest86+ missing in
GRUB2" (bnc#780622), as in memtest86+'s package installation calls
update-bootloader without refresh.