use grub2-probe to get device map
We get the translated os device to grub device mapping thru the
grub2-probe result. By means of this new method we can avoid the mess of
requirement to create file /boot/grub2/device.map which is intented for
perl bootloader to work but is not necessary for grub2.

This fixes the integration issue with kiwi as it couldn't provide such
device map like yast. (bnc#804065)