better way to know kerenl device of physical drive
In previous patch we use udevmap to know the kernel device of physical
disk drive. It works for most cases but for the mdadm software raid
testing only reports md device and no links to physical device as we
would expect. This commit use partition map's disk field to test by the
grub2-probe to determine the device map.