update-bootloader refresh cleanup and introduce reinit
The --refresh is supposed to update the config, it uses avoid_init=0 to
communicate bootloader modules and will enforce the reinstall. This is
not optimal for modern bootloaders like grub as they don't require
reinstall to make new config effective.

The refresh method is modified to use avoid_init=1 and keep avod_init=0
for lilo. However we still have the need to install bootloader for
circumstance like bootloader package update, thus we introduce a new
reinit option to serve that purpose which exactly uses avoid_init=0.