Fix installation boot parameters not written
When uefi secure boot enabled, the linuxefi command will be used to load
and verify kernel image signature. The grub.cfg parsing has to match
this command in order to return image section information structure to
yast, if not the append will not found in the image section and will
lost if writing to disk. (bnc#824609)