fix update
ListFiles is intented to list all possible configuration files, so it
can be backuped and also allow single commit update. Implementation
based on existence of file breaks update process which create at first
<file>.new file and then copy it to target location, but when upgrade
from grub1 it create only new files and never copy it to target location
as it is not listed as grub2 cfg files.