added support for modifying default boot options
I've re-arranged the package layout a bit. There's now a 'pbl' tool with
its own set of options.

update-bootloader and bootloader_entry are just symlinks to it.

'pbl --ACTION' basically maps to running /usr/lib/bootloader/<LOADER>/ACTION.
Some pbl options need an argument. If so, it's passed as argument to the
ACTION script.