grub2-efi: Allow fallback to removable booting
EFI support 2 default ways of booting. It can either store a variable
in nvram that points to an EFI binary or it can use a default efi binary
location in the "removable media" fallback case.

When we have no nvram to store things to, default to the removable media
case. That way we can boot on EFI systems that don't have efi vars (like
the new U-Boot based EFI implementation).

Signed-off-by: Alexander Graf <agraf@suse.de>