add --get-option to pbl (bsc #1033776, bsc #1050349)
This option returns the boot option as set explicitly in the boot loader
configuration (e.g. /etc/default/grub).

It does not try to provide any auto-generated boot options (like 'root' for
example).

It returns the result in the form 'foo=bar' or 'foo' depending on whether it
was an option with argument or without. It returns nothing if the option was
not set.