grub2/install: reset error code when passing through recover code
The script tries to identify the device, 'fails', sets err=1 and
then goes into an error handler to recover from the situation.
In this case though, irrespective of the result of the error handler,
the final result remained err=1.

Reset err to 0 before attempting to write the config on the
auto-detected device.

https://bugzilla.opensuse.org/show_bug.cgi?id=1198197